mcp_host/host_les2/requirements.txt
playwright==1.27.0
python-dotenv==0.21.0
